refactor: restructure App component and separate HomePage and ProductDetail into dedicated files
- Removed inline HomePage and ProductDetail components from App.jsx for better organization.
- Introduced new HomePage and ProductDetail components in their respective files.
- Simplified Header component by integrating hooks directly, removing unnecessary props.
- Updated routing to reflect the new component structure, enhancing maintainability and readability.
